[stm32f1] More robust workaround for I2C errata
The stm32f1 has an issue that can cause the i2c Lines
to be "stuck". This patch implements a more thorough
version of the errata fix, and adds a retry in case
of bus busy at every attempt to enter master mode.
It also passes i2c errors more reliably to upstream
code (which will call shutdown, but still).

Signed-off-by: Vladimir Vukicevic <vladimir@pobox.com>